Made changes to Javadoc, separated presenter functionality from the searchcontroller and the corresponding changes.
