+++ src/main/java/org/hydev/mcpm/client/commands/controllers/SearchPackagesController.java
@@ -4,6 +4,7 @@ import org.hydev.mcpm.client.display.presenters.Table;
 import org.hydev.mcpm.client.models.PluginModel;
 import org.hydev.mcpm.client.search.SearchPackagesBoundary;
 import org.hydev.mcpm.client.search.SearchPackagesInput;
+import org.hydev.mcpm.client.search.SearchPackagesResult;
 import org.hydev.mcpm.client.search.SearchPackagesType;
 
 import javax.annotation.Nullable;
@@ -14,9 +15,8 @@ import java.util.function.Consumer;
 /**
  * Handles the user input for a search.
  *
- * @author Jerry
  */
-public record SearchPackagesController(SearchPackagesBoundary searcher, @Nullable PageBoundary pageController) {
+public record SearchPackagesController(SearchPackagesBoundary searcher) {
 
     /**
      * Load plugins and output status to log.
@@ -24,33 +24,14 @@ public record SearchPackagesController(SearchPackagesBoundary searcher, @Nullabl
      * @param type String that specifies the type of search.
      * @param text String that specifies the search text.
      * @param noCache Specifies whether to use local cache or not.
-     * @param log Callback for status for log events.
+     *
+     * @return The search result.
      */
-    public void searchPackages(String type, String text, boolean noCache, Consumer<String> log) {
+    public SearchPackagesResult searchPackages(String type, String text, boolean noCache) {
         SearchPackagesInput inp = new SearchPackagesInput(
                 SearchPackagesType.valueOf("BY_" + type.toUpperCase()),
                 text,
                 noCache);
-        var result = searcher.search(inp);
-        log.accept("Search State: " + result.state().toString());
-
-        // Print the plugins found
-        var list = result.plugins().stream()
-            .map(PluginModel::getLatestPluginVersion)
-            .filter(Optional::isPresent)
-            .map(it -> it.get().meta())
-            .toList();
-
-        var table = new Table(List.of(":Name", "Author", "Version:"),
-                list.stream().map(p -> List.of("&a" + p.name(), "&e" + p.getFirstAuthor(), p.version())).toList());
-
-        // Pagination
-        if (pageController != null) {
-            pageController.store(table);
-            log.accept(pageController.formatPage(1));
-        }
-        else {
-            log.accept(table.toString());
-        }
+        return searcher.search(inp);
     }
 }
